「B2001 入门测试题目」更改代码

+++ B2001 入门测试题目/main.cpp
@@ -4,9 +4,9 @@ using namespace std;
 int main(){
     // cout<<"请输入一个两位数";
     // 注释这行的原因是可能影响自动改卷机
-    int a,b;
+    long a,b;
     cin>>a>>b;
-    const int c = a+b;
+    const long c = a+b;
     cout<<c;
     
     return 0;
